react changes with babel-react preset
