+++ babel.config.js
@@ -1,3 +1,6 @@
 module.exports = {
-  presets: ["@babel/preset-env"],
+  presets: [
+    "@babel/preset-env",
+    ["@babel/preset-react", { runtime: "automatic" }],
+  ],
 };
