+++ src/index.js
@@ -1,13 +1,3 @@
-import "./styles/index.scss";
-const elvenRecipe = {
-  strips: 2,
-  ironIgnot: 1,
-  refine: 4,
-};
-const sample = {
-  ...elvenRecipe,
-  qty: 23,
-  store: "costcos",
-  address: "unavailables",
-};
-console.log(elvenRecipe, sample, "hellp");
+import { render } from "react-dom";
+import App from "./components/app";
+render(<App />, document.getElementById("root"));
